refactor(#528): deconstructs middleware module into components
